move algolia init logic to wrapper component
