+++ website/src/theme/SearchBar/index.js
@@ -1,63 +1,9 @@
-/**
- * Copyright (c) 2017-present, Facebook, Inc.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import React, { useRef, useCallback } from 'react';
+import React from 'react';
 import classnames from 'classnames';
+import DocSearch from '../../components/DocSearch';
 
-import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
-import { useHistory } from '@docusaurus/router';
-
-const Search = props => {
-  const initialized = useRef(false);
-  const searchBarRef = useRef(null);
-  const { siteConfig = {} } = useDocusaurusContext();
-  const {
-    themeConfig: { algolia },
-  } = siteConfig;
-  const history = useHistory();
-
-  const initAlgolia = () => {
-    if (!initialized.current) {
-      window.docsearch({
-        appId: algolia.appId,
-        apiKey: algolia.apiKey,
-        indexName: algolia.indexName,
-        inputSelector: '#search_input_react',
-        algoliaOptions: algolia.algoliaOptions,
-        // Override algolia's default selection event, allowing us to do client-side
-        // navigation and avoiding a full page refresh.
-        handleSelected: (_input, _event, suggestion) => {
-          // Use an anchor tag to parse the absolute url into a relative url
-          // Alternatively, we can use new URL(suggestion.url) but its not supported in IE
-          const a = document.createElement('a');
-          a.href = suggestion.url;
-
-          // Algolia use closest parent element id #__docusaurus when a h1 page title does not have an id
-          // So, we can safely remove it. See https://github.com/facebook/docusaurus/issues/1828 for more details.
-          const routePath = `#__docusaurus` === a.hash ? `${a.pathname}` : `${a.pathname}${a.hash}`;
-          history.push(routePath);
-        },
-      });
-      initialized.current = true;
-    }
-  };
-
-  const toggleSearchIconClick = useCallback(
-    e => {
-      if (!searchBarRef.current.contains(e.target)) {
-        searchBarRef.current.focus();
-      }
-
-      props.handleSearchBarToggle(!props.isSearchBarExpanded);
-    },
-    [props.isSearchBarExpanded],
-  );
-
-  return (
+const Search = props => (
+  <DocSearch searchElId="#search_input_react">
     <div className="navbar__search" key="search-box">
       <span
         aria-label="expand searchbar"
@@ -65,8 +11,6 @@ const Search = props => {
         className={classnames('search-icon', {
           'search-icon-hidden': props.isSearchBarExpanded,
         })}
-        onClick={toggleSearchIconClick}
-        onKeyDown={toggleSearchIconClick}
         tabIndex={0}
       />
       <input
@@ -79,14 +23,9 @@ const Search = props => {
           { 'search-bar-expanded': props.isSearchBarExpanded },
           { 'search-bar': !props.isSearchBarExpanded },
         )}
-        onClick={initAlgolia}
-        onMouseOver={initAlgolia}
-        onFocus={toggleSearchIconClick}
-        onBlur={toggleSearchIconClick}
-        ref={searchBarRef}
       />
     </div>
-  );
-};
+  </DocSearch>
+);
 
 export default Search;
